L11T02_CountNonDivisible._v01.c / Task / Analysis
/*
L11T02_CountNonDivisible.c

Calculate the number of elements of an array that are not divisors of each element.
Programming language:  
You are given a non-empty zero-indexed array A consisting of N integers.

For each number A[i] such that 0 ≤ i < N, we want to count the number of elements of the array that are not the divisors of A[i]. We say that these elements are non-divisors.

For example, consider integer N = 5 and array A such that:

    A[0] = 3
    A[1] = 1
    A[2] = 2
    A[3] = 3
    A[4] = 6
For the following elements:

A[0] = 3, the non-divisors are: 2, 6,
A[1] = 1, the non-divisors are: 3, 2, 3, 6,
A[2] = 2, the non-divisors are: 3, 3, 6,
A[3] = 3, the non-divisors are: 2, 6,
A[4] = 6, there aren't any non-divisors.
Assume that the following declarations are given:

struct Results {
  int * C;
  int L;
};
Write a function:

struct Results solution(int A[], int N);
that, given a non-empty zero-indexed array A consisting of N integers, returns a sequence of integers representing the amount of non-divisors.

The sequence should be returned as:

a structure Results (in C), or
a vector of integers (in C++), or
a record Results (in Pascal), or
an array of integers (in any other programming language).
For example, given:

    A[0] = 3
    A[1] = 1
    A[2] = 2
    A[3] = 3
    A[4] = 6
the function should return [2, 4, 3, 2, 0], as explained above.

Assume that:

N is an integer within the range [1..50,000];
each element of array A is an integer within the range [1..2 * N].
Complexity:

expected worst-case time complexity is O(N*log(N));
expected worst-case space complexity is O(N), beyond input storage (not counting the storage required for input arguments).
Elements of input arrays can be modified.
Copyright 2009–2017 by Codility Limited. All Rights Reserved. Unauthorized copying, publication or disclosure prohibited.

https://codility.com/programmers/lessons/11-sieve_of_eratosthenes/count_non_divisible/


Correctness	Performance	Task score


	
*/



/*
18:40 start:

A[0] = 3 : 3/3; 3/1; 2/3; 3/3; 6/3;
A[1] = 1 : 1/3; 1/1; 1/2; 1/3; 1/6; 
A[2] = 2 : 2/3; 2/1; 2/2; 2/3; 2/6
A[3] = 3
A[4] = 6

non-divisor: A[i] % A[i] != 0
if A[i] % A[i] != 0
	non_divisor++
*(result.C+result.L) = non_divisor
result.L++

18:54 run test, passed. submit:

Correctness	Performance	Task score
100% 25%	66%

Performance tests
▶ medium_random 
medium, random numbers length = 5,000 ✔OK
▶ large_range 
1, 2, ..., N, length = ~20,000 ✘TIMEOUT ERROR 
running time: 1.18 sec., time limit: 0.10 sec.
▶ large_random 
large, random numbers, length = ~30,000 ✘TIMEOUT ERROR 
running time: 2.64 sec., time limit: 0.10 sec.
▶ large_extreme 
large, all the same values, length = 50,000 ✘TIMEOUT ERROR 
running time: >6.00 sec., time limit: 0.10 sec.
	
*/