L12_Euclidean_algorithm
Chapter 12
Euclidean algorithm
The Euclidean algorithm is one of the oldest numerical algorithms still to be in common
use. It solves the problem of computing the greatest common divisor (gcd) of two positive
integers.
12.1. Euclidean algorithm by subtraction
The original version of Euclid’s algorithm is based on subtraction: we recursively subtract
the smaller number from the larger.
12.1: Greatest common divisor by subtraction.
1 def gcd(a, b):
2 if a == b:
3 return a
4 if a > b:
5 gcd(a - b, b)
6 else:
7 gcd(a, b - a)
Let’s estimate this algorithm’s time complexity (based on n = a+b). The number of steps can
be linear, for e.g. gcd(x, 1), so the time complexity is O(n). This is the worst-case complexity,
because the value x + y decreases with every step.
12.2. Euclidean algorithm by division
Let’s start by understanding the algorithm and then go on to prove its correctness. For two
given numbers a and b, such that a ˇ b:
• if b | a, then gcd(a, b) = b,
• otherwise gcd(a, b) = gcd(b, a mod b).
gcd(24, 9) = gcd(9, 6) = gcd(6, 3) = 3
24 mod 9 = 6 9 mod 6 = 3 6 mod 3 = 0
•c Copyright 2015 by Codility Limited. All Rights Reserved. Unauthorized copying or publication prohibited.
1
Let’s prove that gcd(a, b) = gcd(b, r), where r = a mod b and a = b · t + r:
• Firstly, let d = gcd(a, b). We get d | (b · t + r) and d | b, so d | r.
Therefore we get gcd(a, b) | gcd(b, r).
• Secondly, let c = gcd(b, r). We get c | b and c | r, so c | a.
Therefore we get gcd(b, r) | gcd(a, b).
Hence gcd(a, b) = gcd(b, r). Notice that we can recursively call a function while a is not
divisible by b.
12.2: Greatest common divisor by dividing.
1 def gcd(a, b):
2 if a % b == 0:
3 return b
4 else:
5 return gcd(b, a % b)
Denote by (ai, bi) pairs of values a and b, for which the above algorithm performs i steps.
Then bi ˇ Fibi≠1 (where Fibi is the i-th Fibonacci number). Inductive proof:
1. for one step, b1 = 0,
2. for two steps, b ˇ 1,
3. for more steps, (ak+1, bk+1) æ (ak, bk) æ (ak≠1, bk≠1), then ak = bk+1, ak≠1 = bk,
bk≠1 = ak mod bk, so ak = q · bk + bk≠1 for some q ˇ 1, so bk+1 ˇ bk + bk≠1.
Fibonacci numbers can be approximated by:
Fibn ¥ ( 1+Ô5
2 )n
Ô5 (12.1)
Thus, the time complexity is logarithmic based on the sum of a and b — O(log(a + b)).
12.3. Binary Euclidean algorithm
This algorithm finds the gcd using only subtraction, binary representation, shifting and parity
testing. We will use a divide and conquer technique.
The following function calculate gcd(a, b, res) = gcd(a, b, 1) · res. So to calculate
gcd(a, b) it suces
to call gcd(a, b, 1) = gcd(a, b).
12.3: Greatest common divisor using binary Euclidean algorithm.
1 def gcd(a, b, res):
2 if a == b:
3 return res * a
4 elif (a % 2 == 0) and (b % 2 == 0):
5 return gcd(a // 2, b // 2, 2 * res)
6 elif (a % 2 == 0):
7 return gcd(a // 2, b, res)
8 elif (b % 2 == 0):
9 return gcd(a, b // 2, res)
10 elif a > b:
11 return gcd(a - b, b, res)
12 else:
13 return gcd(a, b - a, res)
2
This algorithm is superior to the previous one for very large integers when it cannot be
assumed that all the arithmetic operations used here can be done in a constant time. Due
to the binary representation, operations are performed in linear time based on the length of
the binary representation, even for very big integers. On the other hand, modulo applied in
algorithm 10.2 has worse time complexity. It exceeds O(log n · log log n), where n = a + b.
Denote by (ai, bi) pairs of values a and b, for which the above algorithm performs i steps.
We have ai+1 ˇ ai, bi+1 ˇ bi, b1 = a1 > 0. In the first three cases, ai+1 · bi+1 ˇ 2 · ai · bi. In the
fourth case, ai+1 · bi+1 ˇ 2 · ai≠1 · bi≠1, because a dierence
of two odd numbers is an even
number. By induction we get:
ai · bi ˇ 2Â i≠1
2 Ê (12.2)
Thus, the time complexity is O(log(a · b)) = O(log a + b) = O(log n). And for very large
integers, O((log n)2), since each arithmetic operation can be done in O(log n) time.
12.4. Least common multiple
The least common multiple (lcm) of two integers a and b is the smallest positive integer that
is divisible by both a and b. There is the following relation:
lcm(a, b) = a·b
gcd(a,b)
Knowing how to compute the gcd(a, b) in O(log(a+b)) time, we can also compute the lcm(a, b)
in the same time complexity.
12.5. Exercise
Problem: Michael, Mark and Matthew collect coins of consecutive face values a, b and c
(each boy has only one kind of coins). The boys have to find the minimum amount of money
that each of them may spend by using only their own coins.
Solution: It is easy to note that we want to find the least common multiple of the three
integers, i.e. lcm(a, b, c). The problem can be generalized for the lcm of exactly n integers.
There is the following relation:
lcm(a1, a2,...,an) = lcm(a1, lcm(a2, a3,...,an))
We simply find the lcm n times, and each step works in logarithmic time.